PatchRes/requirements.txt
faiss_cpu==1.8.0.post1
matplotlib==3.7.0
numpy==1.24.4
opencv_python==4.10.0.84
Pillow==9.0.0
scikit_learn==0.24.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
